voyager_helpers gem, location logic
