--- voyager_helpers/.gitignore
@@ -1,16 +1,0 @@
-/.bundle/
-/.yardoc
-/Gemfile.lock
-/_yardoc/
-/coverage/
-/doc/
-/pkg/
-/spec/reports/
-/tmp/
-*.bundle
-*.so
-*.o
-*.a
-mkmf.log
-/lib/debug.rb
-/lib/run_at_5_10.rb
