--- voyager_helpers/.rspec
@@ -1,3 +1,0 @@
---color
---require spec_helper
---format documentation
