--- voyager_helpers/Gemfile
@@ -1,8 +1,0 @@
-source 'https://rubygems.org'
-
-# Specify your gem's dependencies in marc_liberation.gemspec
-gemspec
-
-group :development, :production do
-  gem 'ruby-oci8', '~> 2.2.1'
-end
