--- voyager_helpers/Rakefile
@@ -1,12 +1,0 @@
-#!/usr/bin/env rake
-
-require "bundler/gem_tasks"
-require 'rspec/core/rake_task'
-
-RSpec::Core::RakeTask.new(:spec)
-
-task :ci do
-  Rake::Task['spec'].invoke
-end
-
-task default: :ci
