--- voyager_helpers/VERSION
@@ -1,1 +1,0 @@
-0.0.1.dev
