--- voyager_helpers/spec/fixtures/earlier_id_dump.txt
@@ -1,6 +1,0 @@
-1 2000-06-08T00:00:00-05:00 2012-05-21T11:40:38-05:00
-2 2000-06-08T00:00:00-05:00 2010-06-16T15:55:32-05:00
-3 2000-06-08T00:00:00-05:00
-4 2000-06-08T00:00:00-05:00 2011-05-04T10:14:46-05:00
-5 2000-06-08T00:00:00-05:00 2013-04-12T13:40:57-05:00
-8 2000-06-08T00:00:00-05:00 2012-07-06T10:21:00-05:00
